Getting containers working
Passing URL to the Brewski API as an environment variable to the web
docker container. Since the URL must be accessible from a web browser,
we can’t just use the internal name or ip address of the api container.

Renamed index.js to app.js, since index.js seems to me to indicate that
it’s being served from a web server (which it isn’t).